Safari-611.1.10/LayoutTests/crypto/workers/subtle/ecdh-import-key-derive-bits-expected.txt

[Worker] Test ECDH deriveBits operation with imported base key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ecdh-import-key-derive-bits.js
PASS [Worker] bytesToHexString(derivedKey) is expectedDerivedKey
PASS successfullyParsed is true

TEST COMPLETE
